Incorporated the review comments. Code refactoring.
