+++ cluster/src/dunit/scala/org/apache/spark/sql/SmartConnectorFunctions.scala
@@ -131,66 +131,60 @@ object SmartConnectorFunctions {
     " {fn FLOOR((-1 * {fn DAYOFYEAR(O_ORDERDATE)} - 1))}, O_ORDERDATE)}" +
     " from orders where O_ORDERKEY = 32"
 
-  private def verifyRowTablePruning(snc: SnappyContext): Unit = {
+  def verifyRowTablePruning(snc: SnappyContext): Unit = {
 
-    val earlierValue = io.snappydata.Property.ColumnBatchSize.get(snc.sessionState.conf)
-    try {
-      io.snappydata.Property.ColumnBatchSize.set(snc.sessionState.conf, "1000")
-      val tpchDataPath = TPCHColumnPartitionedTable.getClass.getResource("/TPCH").getPath
-      val buckets_Order_Lineitem = "5"
-      TPCHColumnPartitionedTable.createPopulateOrderTable(snc, tpchDataPath,
-        isSnappy = true, buckets_Order_Lineitem, null, provider = "row")
+    val tpchDataPath = TPCHColumnPartitionedTable.getClass.getResource("/TPCH").getPath
+    val buckets_Order_Lineitem = "5"
+    TPCHColumnPartitionedTable.createPopulateOrderTable(snc, tpchDataPath,
+      isSnappy = true, buckets_Order_Lineitem, null, provider = "row")
 
-      def validateSinglePartition(df: DataFrame, bucketId: Int): Unit = {
-        val plan = df.queryExecution.executedPlan.collectFirst {
-          case c: RowTableScan => c
-        }
-
-        val scanRDD = plan.map(_.dataRDD).
-          getOrElse(throw new AssertionError("Expecting RowTable Scan"))
-        val partitions = scanRDD.partitions
-        assert(plan.get.outputPartitioning == SinglePartition)
-        assert(partitions.length == 1, {
-          val sb = new StringBuilder("Pruning not in effect ? partitions found ")
-          partitions.foreach(p => sb.append(p.index).append(","))
-          sb.toString
-        })
-        val bstr = partitions(0) match {
-          case zp: ZippedPartitionsPartition => zp.partitionValues.map {
-            case mb: MultiBucketExecutorPartition => mb.bucketsString
-          }
-          case _ => Nil
-        }
-
-        // each BucketExecutor must have only one bucket.
-        // there are 2 BucketExecutor entries due to ZipPartion of RowBuffer.
-        assert(bstr.forall(_.toInt == bucketId), s"Expected $bucketId, found $bstr")
+    def validateSinglePartition(df: DataFrame, bucketId: Int): Unit = {
+      val plan = df.queryExecution.executedPlan.collectFirst {
+        case c: RowTableScan => c
       }
 
-      validateSinglePartition(executeQuery(snc, query1 + 1, 1), 4)
-      validateSinglePartition(executeQuery(snc, query1 + 32, 32), 0)
-      validateSinglePartition(executeQuery(snc, query1 + 801, 801), 4)
-      // repeating the query deliberately
-      validateSinglePartition(executeQuery(snc, query1 + 801, 801), 4)
-      validateSinglePartition(executeQuery(snc, query1 + 1408, 1408), 0)
-      validateSinglePartition(executeQuery(snc, query1 + 1409, 1409), 2)
-      validateSinglePartition(executeQuery(snc, query1 + 1410, 1410), 0)
-      validateSinglePartition(executeQuery(snc, query1 + 1796, 1796), 4)
-      validateSinglePartition(executeQuery(snc, query1 + 801, 801), 4)
-      executeQuery(snc, query1 + "'1'", 1)
-      executeQuery(snc, query1 + "'32'", 32)
-      executeQuery(snc, query2, 1)
-      executeQuery(snc, query3, 801)
-      executeQuery(snc, query4, 801)
-      executeQuery(snc, query5, 1410)
+      val scanRDD = plan.map(_.dataRDD).
+        getOrElse(throw new AssertionError("Expecting RowTable Scan"))
+      val partitions = scanRDD.partitions
+      assert(plan.get.outputPartitioning == SinglePartition)
+      assert(partitions.length == 1, {
+        val sb = new StringBuilder("Pruning not in effect ? partitions found ")
+        partitions.foreach(p => sb.append(p.index).append(","))
+        sb.toString
+      })
+      val bstr = partitions(0) match {
+        case zp: ZippedPartitionsPartition => zp.partitionValues.map {
+          case mb: MultiBucketExecutorPartition => mb.bucketsString
+        }
+        case _ => Nil
+      }
 
-      val df = executeQuery(snc, query6, 32, false)
-      val r = df.collect()(0)
-      assert(r.getDate(0).toString.equals("1995-07-16"))
-      assert(r.getDate(1).toString.equals("1994-12-30"))
-    } finally {
-      io.snappydata.Property.ColumnBatchSize.set(snc.sessionState.conf, earlierValue)
+      // each BucketExecutor must have only one bucket.
+      // there are 2 BucketExecutor entries due to ZipPartion of RowBuffer.
+      assert(bstr.forall(_.toInt == bucketId), s"Expected $bucketId, found $bstr")
     }
+
+    validateSinglePartition(executeQuery(snc, query1 + 1, 1), 4)
+    validateSinglePartition(executeQuery(snc, query1 + 32, 32), 0)
+    validateSinglePartition(executeQuery(snc, query1 + 801, 801), 4)
+    // repeating the query deliberately
+    validateSinglePartition(executeQuery(snc, query1 + 801, 801), 4)
+    validateSinglePartition(executeQuery(snc, query1 + 1408, 1408), 0)
+    validateSinglePartition(executeQuery(snc, query1 + 1409, 1409), 2)
+    validateSinglePartition(executeQuery(snc, query1 + 1410, 1410), 0)
+    validateSinglePartition(executeQuery(snc, query1 + 1796, 1796), 4)
+    validateSinglePartition(executeQuery(snc, query1 + 801, 801), 4)
+    executeQuery(snc, query1 + "'1'", 1)
+    executeQuery(snc, query1 + "'32'", 32)
+    executeQuery(snc, query2, 1)
+    executeQuery(snc, query3, 801)
+    executeQuery(snc, query4, 801)
+    executeQuery(snc, query5, 1410)
+
+    val df = executeQuery(snc, query6, 32, false)
+    val r = df.collect()(0)
+    assert(r.getDate(0).toString.equals("1995-07-16"))
+    assert(r.getDate(1).toString.equals("1994-12-30"))
   }
 
   private def executeQuery(snc: SnappyContext, sql: String, orderKey: Int,
