+++ core/src/main/scala/org/apache/spark/sql/execution/columnar/impl/JDBCSourceAsColumnarStore.scala
@@ -674,15 +674,7 @@ final class ColumnarStorePartitionedRDD(
             region.asInstanceOf[PartitionedRegion])
           allPartitions
         } else {
-          val pr = region.asInstanceOf[PartitionedRegion]
-          val distMembers = StoreUtils.getBucketOwnersForRead(bucketId, pr)
-          val prefNodes = new ArrayBuffer[String](2)
-          distMembers.foreach(m => SnappyContext.getBlockId(m.canonicalString()) match {
-            case Some(b) => prefNodes += Utils.getHostExecutorId(b.blockId)
-            case _ =>
-          })
-          Array(new MultiBucketExecutorPartition(0, ArrayBuffer(bucketId),
-            pr.getTotalNumberOfBuckets, prefNodes))
+          Utils.getPartitions(region, bucketId)
         }
     }
   }
@@ -877,7 +869,7 @@ class SmartConnectorRowRDD(_session: SnappySession,
     _connProperties: ConnectionProperties,
     _filters: Array[Expression],
     _partEval: () => Array[Partition],
-    _partitionPruner: () => Int,
+    _partitionPruner: () => Int = () => -1,
     private var catalogSchemaVersion: Long,
     _commitTx: Boolean, _delayRollover: Boolean)
     extends RowFormatScanRDD(_session, _tableName, _isPartitioned, _columns,
@@ -1003,7 +995,7 @@ class SmartConnectorRowRDD(_session: SnappySession,
     // (updated values in ParamLiteral will take care of updating filters)
     evaluateWhereClause()
     val parts = partitionEvaluator()
-    if(parts.length == _partEval.apply().length){
+    if(parts.length == _partEval().length){
       return getPartitionEvaluator
     }
     logDebug(s"$toString.getPartitions: $tableName partitions ${parts.mkString("; ")}")
@@ -1012,9 +1004,9 @@ class SmartConnectorRowRDD(_session: SnappySession,
 
   def getPartitionEvaluator: Array[Partition] = {
     partitionPruner() match {
-      case -1 => _partEval.apply()
+      case -1 => _partEval()
       case bucketId =>
-        val part = _partEval.apply()(bucketId).asInstanceOf[SmartExecutorBucketPartition]
+        val part = _partEval()(bucketId).asInstanceOf[SmartExecutorBucketPartition]
         Array(new SmartExecutorBucketPartition(0, bucketId, part.hostList))
     }
   }
